stylegan2-tpu/requirements.txt
absl-py==0.7.0
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
gast==0.2.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.19.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
MarkupSafe==1.1.1
mock==2.0.0
numpy==1.16.2
pbr==5.1.2
Pillow==5.4.1
protobuf==3.6.1
requests==2.21.0
six==1.12.0
tensorflow-gpu==1.15.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wget==3.2
